Added documentation and moved base class to bottom of file.
I added quite a bit of documentation, both pod and comments.
I also structured the documentation with the discussion of creating
your own derived plotting classes at the end of the document. It made a
lot of sense to keep the base class code close to that documentation,
which is why I moved that code to the end of the file.