Substantially updated Spike plotting.
Added the option to choose horizontal or vertical lines, and to choose
the 'baseline' of those lines. This impacted the padding calculations, so
I had to add min and max functionality.

In the process, I created and then removed (but retained as comments) some
code to handle generic padding. I will have to decide how to handle generic
padding when I overhaul the min/max algorithm.