Major overhaul of the automatic bounds calculation
I've almost completed the bound checking overhaul. At this point,
I still need to look over the axis code to ensure that it uses the
new code correctly and smartly. I also have lots of stuff yet to
document, and I've discovered that the collation code for plot types
is much harder to write than the old min/max code. :-(