Added save-to-file function and right-click action!
I have been making a number of interactive plots where I have to
install an 'Autoscale' button just to restore autoscaling after
I've been zooming around for a while. The solution to this, in my
opinion, is to create a context-sensitive menu with options like
autoscaling and the like.

I also discovered how to get the rasterized view of a window, which
allows for rudimentary (rasterized) image output. Prima has excellent
support for rasterized file formats, so adding it was surprisingly
easy. I added an option to the context menu for that. :-)