Improved Save As, added Copy
I realized that Prima has clipboard management, so copying the image
to a clipboard shouldn't be too difficult. And, it wasn't! I ran into
trouble testing the clipboard because the menu refused to disappear.
I revised the waiting procedure so that I now use a Prima timer.
Everything is working splendidly now!