Substantially improved the documentation.
I am getting ready to push this library out to the public as an
alpha, which means that at least *something* needs to be well
documented. For now, that something is the Simple interface. I am
also working on the other docs, but I made a huge push to get the
Simple interface documented. At this point, the only thing missing
is the major documentation about using the plot command.