Fixed autoscaling bug with function-only plots.
Function only plots croaked when you attempt to autoscale them.
That is because the widget's compute_min_max_for function did not
properly handle the case when all the data is bad, which is precisely
what the function returns for its x data. This defaults to a zero
pixel-padded width at the current view width.