Switched Palette to use PDL::Drawing::Prima::Utils
PDL::Drawing::Prima now has a strict requirement on Prima, which
breaks the forking behavior on Unixes in the Simple module. Palette
was the only module that had a dependence on PDL::Drawing::Prima,
and that only for the color space manipulation functions. I've
pulled those into a seperate module in PDL::Drawing::Prima::Utils
and modified the Palette dependency accordingly.