Better Autoscaling (and some doc tweaks)
I got sick of the first view of a plot having incorrect spacing
for the y-axis, so I hammered on it. The code is even more
confusing than when I started, but at least it works. At some
point I will need to closely examine the whole auto-scaling code
to figure out how to better structure it. At the moment, it's like
crawling through spaghetti.