Defensive programming, and fixed a degenerate scaling bug.
I found an autoscaing bug when the data was degenerate (only one
point) *and* that one point was negative, not positive. In the
process of tracking it down, I added some defensive programming.
The bug turned out to be how the linear Scaling's
min_max_for_degenerate(). That function properly handled the case
of a zero value, but the degenerate nonzero case involed calculations
that implicitly assumed a positive value; that logic has been
corrected.