Fixed y auto-scaling even when min/max were set.
I found an issue in which the y axis would autoscale even when I
explicitly set the min or max (or both). Turned out that it had to
do with an extra step in compute_min_max_for() that was recently
added in which the y-scaling was determined ahead of time if the
x-scaling was the thing actually being calculated. That helped fix
the fact that auto-scaling should be idempotent, but wasn't. This
additional fix patches the issue that the previous fix introduced.