Properly handle cached values.
The plot widget keeps track of the amount of space needed for the
axis labels and other things. It does this at the end of each
painting operation. However, the space needed is measured in pixels
and are particular to the canvas of interest. Creating a plot on a
different canvas, especially a postscript canvas, can cause trouble
if its sizes are different.

The ultimate solution to this problem is probably to create a
separate function that computes the space needed for tick labels.
Since such a function does not yet exist, the work around implemented
in this commit assumes that the new canvas has dimensions that are
scaled multiples of the widget's dimensions, and passes around the
ratio of the new canvas's height to the widget's height.