Added Anotation dataset and Anotation::Region; needs more docs.
I've wanted the ability to shade various regions of a plot. The
Anotation::Region plot type achieves just that. It lets you fill
an area and allows for highly flexible edge specification.

The documentation needs to be expanded the min/max collation needs
to be implemented, but the basics work for now!