Finally simplified linear scaling tick algorithm.
Watching an animation with tick jitter finally convinced me that
it was time to fix the tick algorithm once and for all. The revised
tick algorithm for linear scaling has the possibility of yielding a
wider range of ticks, but at least the spacing is always consistent.
The range of ticks (sometimes as few as four, sometimes as many as
seven) could be tightened by allowing tick intervals such as 0.4 or
0.8 times the order of magnitude.