Modified pal::Rainbow; added pal::RainbowSV
pal::Rainbow had hues 0-360, but both 0 and 360 are red, and this
led to color conflation between the smallest and largest values.
pal::Rainbow now uses hues 0-300, which is red through purple.

I also added pal::RainbowSV to make less-than-unity saturation and
value settings in a rainbow palette easier to create.