Added a new dialog for setting basic plot properties.
After watching somebody struggle trying to set basic plot properties on a live
plot, I decided that it was finally time to implement a basic dialog box to set
these things. The dialog can be acivated by the user by selecting the
Properties entry in the pop-up menu. It can be programmatically invoked by
the method set_propeties_dialog.