Use Prima's new isEPS flag.
For Mac LaTeX rendering and Mac Preview, the postscript images need
to have a special comment to indicate that they are EPS files. Otherwise
they simply do not render properly. Current development versions of
Prima (and Prima v1.40 when it is released) have support for a new
flag, isEPS, which properly modifies the postscript output to give
something that Mac's Preview and latex properly render and handle.