Enhancements to the Save as Postscript dialog
With the isEPS changes already committed, it now makes sense to
have the Save as Postscript dialog use the file extension .eps and
to refer to these files as Encapsulated Postscript. Also, this
commit appends a .eps extension to file names that do not have
that extension.

It occurred to me that if the user types a file name without the
extension and the file name with the extension exists, they will
*not* be prompted to overwrite, because the default file extension
will be added after the dialog closes. This needs to be fixed, but
is probably better fixed with a general Save As... overhaul.