Improve EPS output
The height was being set incorrectly. I'm not sure what it was doing
before, but now it uses method calls instead of hash lookups, which
appears to be the proper way of obtaining a font height.

For postscript output, this now restricts to using device fonts only.
Otherwise, the GUI font is rasterized and drawn pixel-by-pixel on the
postscrip canvas, and it looks terrible.