Add default_save_dir, default_save_format, and ini handling
I got tired of always navigating to a subfolder and picking the
same (but not default) file extension for my figures. So, I added
two properties to a Prima::Plot, named above, which are worked in
to the save dialog.

I also added functionality to the simple interface to read
prima-simple.ini and add any key/value pairs to the plot creation
command. This way I can specify the default image information in
a single file, outside of my various analysis scripts.