Fix subtle bug in Grid's guess_scaling_for
The function tried to determine whether the spacing for a collection
of bin centers was linear or logarithmic. The calculation for
log scaling ussed division. If the bin centers have the value of
zero, this can lead to problems later when trying to compare the
linear score to the not-a-number log score. The code now explicitly
checks for a zero element and returns linear in that case.

I feel like there has been a subtle shift in how "the system"
processes inf. I'm not sure if it's PDL or my OS or hardware, but
I feel that the calculation should say 4/0 = inf, and propogate that
inf through the rest of the calculations. At some point, though, the
inf turns into NaN, and things broke down.