Add capability for rendering color maps!
It is now possible to render color maps. Previously, when rendering
a heat map, for example, one would have used a matrix plot type
along with a palette. Such palettes are now referred to as color
maps. It is now possible to specify the color map among the dataset
parameters, or AMONG THE PLOT'S PARAMETERS. If utilizing the latter,
the color map is automatically rendered on the right side of the
plot.