Keep image export from crashing live windows.
Palette drawing changes the palette's notion of its "widget" to
the given canvas so that it can recompute distances. However, when
done, the palette continues to hold a reference to the old canvas,
so that window redrawing crashes. This fixes the situation by
backing up and restoring the existing widget at the start and end
of drawing.

This feels a bit ugly. The reason I change the widget when drawing
is so that the SizeSpec objects are properly set up. It seems that
there should be a smarter way to handle SizeSpec objects that does
not involve changing the palette's notion of its widget.