Improve solution from d09ad8
The aforementioned commit tried to fix a bug but did it in a bad
way. Looking carefully at the bug suggested that I needed a
different approach, which this commit implements. Now SizeSpec
objects for a widget are returned in a hashref. If the draw is
being implemented on something other than the default widget, a
new hashref is constructed for that drawable and returned instead.